+++ OpenEstate-IS24-REST-webapp/src/main/java/org/openestate/is24/restapi/webapp/VerificationServlet.java
@@ -29,7 +29,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import oauth.signpost.exception.OAuthException;
 import org.apache.commons.io.IOExceptionWithCause;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.openestate.is24.restapi.AbstractClient;
 import org.openestate.is24.restapi.DefaultClient;
 import org.openestate.is24.restapi.utils.Authorization;