+++ OpenEstate-IS24-REST/src/main/java/org/openestate/is24/restapi/DefaultClient.java
@@ -25,8 +25,8 @@ import java.net.URL;
 import java.util.zip.GZIPInputStream;
 import oauth.signpost.exception.OAuthException;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang.RandomStringUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.openestate.is24.restapi.utils.RequestMethod;
 import org.openestate.is24.restapi.utils.Response;