+++ OpenEstate-IS24-REST/src/main/java/org/openestate/is24/restapi/utils/ExportHandler.java
@@ -32,7 +32,7 @@ import javax.xml.bind.JAXBException;
 import oauth.signpost.exception.OAuthException;
 import org.apache.commons.io.IOExceptionWithCause;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.openestate.is24.restapi.AbstractClient;
 import org.openestate.is24.restapi.ImportExport;
 import org.openestate.is24.restapi.xml.common.Attachment;