+++ OpenEstate-IS24-REST/src/main/java/org/openestate/is24/restapi/utils/ExportMessage.java
@@ -16,7 +16,7 @@
 
 package org.openestate.is24.restapi.utils;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 /**
  * A message, that was returned during a bulk export.