+++ OpenEstate-IS24-REST/src/main/java/org/openestate/is24/restapi/utils/RequestFailedException.java
@@ -16,7 +16,7 @@
 
 package org.openestate.is24.restapi.utils;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.openestate.is24.restapi.xml.common.Message;
 import org.openestate.is24.restapi.xml.common.Messages;
 import org.slf4j.Logger;